Changed more details in migrations

+++ database/migrations/2020_11_16_085416_create_guides_table.php
@@ -17,7 +17,7 @@ class CreateGuidesTable extends Migration
             $table->bigIncrements('id');
             $table->string('title');
             $table->string('blog');
-            $table->integer('guide_likes_id');
+            $table->unsignedBigInteger('guide_likes_id');
             $table->timestamps();
         });
     }
